remove fp manager
